Added iMommy pages
will have to test the video tag or find other ways to code the videos for tutorials, which is mostly pseudocode/comments right now. was also thinking of a mom profile picture where users can insert a pic of their mom on the home page! what do you guys think? but we could focus on the main functions first.